+++ testapp/tests/test_e2e_selectize.py
@@ -98,20 +98,20 @@ def form(view):
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', views.keys())
 def test_form_validated(page, form, viewname):
-    assert page.query_selector('django-formset form') is not None
+    expect(page.locator('django-formset form')).to_have_count(1)
     if form.name in ['selection_required', 'static_selection', 'multi_selection']:
-        assert page.query_selector('django-formset form:valid') is None
-        assert page.query_selector('django-formset form:invalid') is not None
+        expect(page.locator('django-formset form:valid')).to_have_count(0)
+        expect(page.locator('django-formset form:invalid')).to_have_count(1)
     else:
-        assert page.query_selector('django-formset form:valid') is not None
-        assert page.query_selector('django-formset form:invalid') is None
+        expect(page.locator('django-formset form:valid')).to_have_count(1)
+        expect(page.locator('django-formset form:invalid')).to_have_count(0)
 
 
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', views.keys())
 def test_initial_value(page, form, viewname):
-    select_element = page.query_selector('django-formset select[is="django-selectize"]')
-    assert select_element is not None
+    select_element = page.locator('django-formset select[is="django-selectize"]')
+    expect(select_element).to_have_count(1)
     if form.name in ['selection_initialized', 'selection_required_initialized']:
         value = select_element.evaluate('elem => elem.value')
         assert value == str(get_initial_opinion().id)
@@ -125,44 +125,42 @@ def test_initial_value(page, form, viewname):
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['selectize1'])
 def test_changing_value(page, form, viewname):
-    input_element = page.query_selector('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
-    assert input_element is not None
-    assert input_element.is_visible()
+    input_element = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
+    expect(input_element).to_be_visible()
     assert input_element.get_attribute('placeholder') == 'Select'
     assert input_element.evaluate('elem => elem.value') == ''
-    field_group_element = page.query_selector('django-formset [role="group"]')
-    assert field_group_element is not None
-    assert 'dj-pristine' in field_group_element.get_attribute('class')
-    assert 'dj-dirty' not in field_group_element.get_attribute('class')
-    dropdown_element = page.query_selector('django-formset .shadow-wrapper .ts-dropdown.single')
-    assert dropdown_element is not None
-    assert dropdown_element.is_hidden()
+    field_group_element = page.locator('django-formset [role="group"]')
+    expect(field_group_element).to_have_count(1)
+    expect(field_group_element).to_have_class('dj-required dj-untouched dj-pristine')
+    expect(field_group_element).not_to_have_class('dj-dirty')
+    dropdown_element = page.locator('django-formset .shadow-wrapper .ts-dropdown.single')
+    expect(dropdown_element).to_have_count(1)
+    expect(dropdown_element).not_to_be_visible()
     input_element.click()
-    assert dropdown_element.is_visible()
-    assert page.query_selector('django-formset form:invalid') is not None
-    pseudo_option = dropdown_element.query_selector('div[data-selectable]:nth-child(9)')
-    assert pseudo_option.is_visible()
+    expect(dropdown_element).to_be_visible()
+    expect(page.locator('django-formset form:invalid')).to_have_count(1)
+    pseudo_option = dropdown_element.locator('div[data-selectable]').nth(8)
+    expect(pseudo_option).to_be_visible()
     initial_opinion = get_initial_opinion()
     assert pseudo_option.get_attribute('data-value') == str(initial_opinion.id)
-    assert pseudo_option.inner_text() == initial_opinion.label
+    expect(pseudo_option).to_have_text(initial_opinion.label)
     pseudo_option.click()
-    assert 'dj-pristine' not in field_group_element.get_attribute('class')
-    assert 'dj-dirty' in field_group_element.get_attribute('class')
-    assert dropdown_element.is_hidden()
-    assert page.query_selector('django-formset form:valid') is not None
-    selected_item_element = page.query_selector('django-formset .shadow-wrapper .ts-wrapper .ts-control div.item')
-    assert selected_item_element is not None
+    expect(field_group_element).to_have_class('dj-required dj-touched dj-dirty')
+    expect(dropdown_element).to_be_hidden()
+    expect(page.locator('django-formset form:valid')).to_have_count(1)
+    selected_item_element = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-control div.item')
+    expect(selected_item_element).to_have_count(1)
     assert selected_item_element.get_attribute('data-value') == str(initial_opinion.id)
-    assert selected_item_element.inner_text() == initial_opinion.label
-    select_element = page.query_selector('django-formset select[is="django-selectize"]')
-    assert select_element is not None
+    expect(selected_item_element).to_have_text(initial_opinion.label)
+    select_element = page.locator('django-formset select[is="django-selectize"]')
+    expect(select_element).to_have_count(1)
     value = select_element.evaluate('elem => elem.value')
     assert value == str(initial_opinion.id)
     input_element.focus()
     page.keyboard.press('Backspace')
     input_element.evaluate('elem => elem.blur()')
-    assert page.query_selector('django-formset .shadow-wrapper .ts-input div.item') is None
-    assert page.query_selector('django-formset form:invalid') is not None
+    expect(page.locator('django-formset .shadow-wrapper .ts-input div.item')).to_have_count(0)
+    expect(page.locator('django-formset form:invalid')).to_have_count(1)
     value = select_element.evaluate('elem => elem.value')
     assert value == ''
 
@@ -170,45 +168,39 @@ def test_changing_value(page, form, viewname):
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['selectize5'])
 def test_add_multiple(page, form, viewname):
-    input_element = page.query_selector('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-multiple"]')
-    assert input_element is not None
-    assert input_element.is_visible()
-    assert input_element.get_attribute('placeholder') == 'Select any'
+    input_element = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-multiple"]')
+    expect(input_element).to_be_visible()
+    expect(input_element).to_have_attribute('placeholder', 'Select any')
     assert input_element.evaluate('elem => elem.value') == ''
-    field_group_element = page.query_selector('django-formset [role="group"]')
-    assert field_group_element is not None
-    assert 'dj-pristine' in field_group_element.get_attribute('class')
-    assert 'dj-untouched' in field_group_element.get_attribute('class')
-    assert 'dj-dirty' not in field_group_element.get_attribute('class')
-    dropdown_element = page.query_selector('django-formset .shadow-wrapper .ts-dropdown.multi')
-    assert dropdown_element is not None
-    assert dropdown_element.is_hidden()
+    field_group_element = page.locator('django-formset [role="group"]')
+    expect(field_group_element).to_have_count(1)
+    expect(field_group_element).to_have_class('dj-required dj-untouched dj-pristine')
+    expect(field_group_element).not_to_have_class('dj-dirty')
+    dropdown_element = page.locator('django-formset .shadow-wrapper .ts-dropdown.multi')
+    expect(dropdown_element).to_have_count(1)
+    expect(dropdown_element).to_be_hidden()
     input_element.click()
-    assert dropdown_element.is_visible()
+    expect(dropdown_element).to_be_visible()
     selected_ids = []
-    pseudo_option = dropdown_element.query_selector('div[data-selectable]:nth-child(3)')
-    selected_ids.append(pseudo_option.get_attribute('data-value'))
-    pseudo_option.click()
-    pseudo_option = dropdown_element.query_selector('div[data-selectable]:nth-child(3)')
-    pseudo_option.click()
-    selected_ids.append(pseudo_option.get_attribute('data-value'))
-    pseudo_option = dropdown_element.query_selector('div[data-selectable]:nth-child(3)')
-    pseudo_option.click()
-    selected_ids.append(pseudo_option.get_attribute('data-value'))
-    assert dropdown_element.is_visible()
+    pseudo_options = dropdown_element.locator('div[data-selectable]')
+    for k in range(3):
+        selected_ids.append(pseudo_options.nth(2).get_attribute('data-value'))
+        pseudo_options.nth(2).click()
+    expect(dropdown_element).to_be_visible()
     input_element.evaluate('elem => elem.blur()')
-    assert dropdown_element.is_hidden()
-    selected_item_elements = page.query_selector_all('django-formset .shadow-wrapper .ts-wrapper .ts-control div.item')
-    assert len(selected_item_elements) == 3
-    assert selected_item_elements[1].get_attribute('data-value') == selected_ids[1]
-    remove_selected_item_element = page.query_selector(f'django-formset .shadow-wrapper .ts-wrapper .ts-control div[data-value="{selected_ids[1]}"].item .remove')
-    assert remove_selected_item_element is not None
+    expect(dropdown_element).to_be_hidden()
+    selected_item_elements = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-control div.item')
+    expect(selected_item_elements).to_have_count(3)
+    for k in range(3):
+        expect(selected_item_elements.nth(k)).to_have_attribute('data-value', selected_ids[k])
+    remove_selected_item_element = page.locator(f'django-formset .shadow-wrapper .ts-wrapper .ts-control div[data-value="{selected_ids[1]}"].item .remove')
+    expect(remove_selected_item_element).to_be_visible()
     remove_selected_item_element.click()
     selected_ids.pop(1)
-    selected_item_elements = page.query_selector_all('django-formset .shadow-wrapper .ts-wrapper .ts-control div.item')
-    assert len(selected_item_elements) == 2
-    select_element = page.query_selector('django-formset select[is="django-selectize"]')
-    assert select_element is not None
+    selected_item_elements = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-control div.item')
+    expect(selected_item_elements).to_have_count(2)
+    select_element = page.locator('django-formset select[is="django-selectize"]')
+    expect(select_element).to_have_count(1)
     values = select_element.evaluate('elem => Array.from(elem.selectedOptions).map(o => o.value)')
     assert set(values) == set(selected_ids)
 
@@ -216,30 +208,28 @@ def test_add_multiple(page, form, viewname):
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['selectize6'])
 def test_change_multiple(page, form, viewname):
-    formset_element = page.query_selector('django-formset')
-    select_element = page.query_selector('django-formset select[is="django-selectize"]')
-    assert select_element is not None
+    formset_element = page.locator('django-formset')
+    select_element = formset_element.locator('select[is="django-selectize"]')
+    expect(select_element).to_be_visible()
     values = select_element.evaluate('elem => Array.from(elem.selectedOptions).map(o => o.value)')
     assert len(values) == 4
     assert set(values) == set(str(i) for i in get_initial_opinions().values_list('id', flat=True))
-    field_group_element = page.query_selector('django-formset [role="group"]')
-    assert field_group_element is not None
-    assert 'dj-pristine' in field_group_element.get_attribute('class')
-    assert 'dj-untouched' in field_group_element.get_attribute('class')
-    assert 'dj-dirty' not in field_group_element.get_attribute('class')
-    remove_selected_item_element = formset_element.query_selector(f'.shadow-wrapper .ts-wrapper .ts-control div[data-value="{values[1]}"].item .remove')
-    assert remove_selected_item_element is not None
+    field_group_element = formset_element.locator('[role="group"]')
+    expect(field_group_element).to_be_visible()
+    expect(field_group_element).to_have_class('dj-untouched dj-pristine')
+    remove_selected_item_element = formset_element.locator(f'.shadow-wrapper .ts-wrapper .ts-control div[data-value="{values[1]}"].item .remove')
+    expect(remove_selected_item_element).to_be_visible()
     remove_selected_item_element.click()
-    item_elements = formset_element.query_selector_all(f'.shadow-wrapper .ts-wrapper .ts-control div.item')
-    assert len(item_elements) == 3
-    assert 'dj-dirty' in field_group_element.get_attribute('class')
+    item_elements = formset_element.locator(f'.shadow-wrapper .ts-wrapper .ts-control div.item')
+    expect(item_elements).to_have_count(3)
+    expect(field_group_element).to_have_class('dj-untouched dj-dirty')
 
 
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['selectize1'])
 def test_lookup_value(page, mocker, form, viewname):
-    input_element = page.query_selector('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
-    assert input_element is not None
+    input_element = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
+    expect(input_element).to_be_visible()
     input_element.click()
     spy = mocker.spy(FormView, 'get')
     page.keyboard.press('1')
@@ -251,33 +241,33 @@ def test_lookup_value(page, mocker, form, viewname):
     content = json.loads(spy.spy_return.content)
     assert content['count'] == 1
     assert content['options'][0]['label'] == "Opinion 0159"
-    dropdown_element = page.query_selector('django-formset .shadow-wrapper .ts-dropdown.single')
-    pseudo_option = dropdown_element.query_selector('div[data-selectable]:nth-child(1)')
-    assert pseudo_option is not None
-    pseudo_option.inner_text() == "Opinion 99"
+    dropdown_element = page.locator('django-formset .shadow-wrapper .ts-dropdown.single')
+    pseudo_option = dropdown_element.locator('div[data-selectable]').nth(0)
+    expect(pseudo_option).to_be_visible()
+    expect(pseudo_option).to_have_text("Opinion 0159")
 
 
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['selectize1'])
 def test_submit_missing(page, view, form, viewname):
-    page.wait_for_selector('django-formset').evaluate('elem => elem.submit()')
-    placeholder_text = page.query_selector('[role="group"] ul.dj-errorlist > li.dj-placeholder').inner_text()
-    assert placeholder_text == Field.default_error_messages['required']
+    page.locator('django-formset').evaluate('elem => elem.submit()')
+    placeholder = page.locator('[role="group"] ul.dj-errorlist > li.dj-placeholder')
+    expect(placeholder).to_have_text(str(Field.default_error_messages['required']))
 
 
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['selectize1'])
 def test_submit_value(page, mocker, view, form, viewname):
-    input_element = page.query_selector('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
-    assert input_element is not None
+    input_element = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
+    expect(input_element).to_be_visible()
     input_element.click()
-    dropdown_element = page.query_selector('django-formset .shadow-wrapper .ts-wrapper .ts-dropdown.single')
-    assert dropdown_element is not None
-    pseudo_option = dropdown_element.query_selector('div[data-selectable]:nth-child(9)')
-    assert pseudo_option is not None
+    dropdown_element = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-dropdown.single')
+    expect(dropdown_element).to_be_visible()
+    pseudo_option = dropdown_element.locator('div[data-selectable]').nth(8)
+    expect(pseudo_option).to_be_visible()
     pseudo_option.click()
     spy = mocker.spy(view.view_class, 'post')
-    page.wait_for_selector('django-formset').evaluate('elem => elem.submit()')
+    page.locator('django-formset').evaluate('elem => elem.submit()')
     request = json.loads(spy.call_args.args[1].body)
     assert request['formset_data']['model_choice'] == str(get_initial_opinion().id)
     assert spy.spy_return.status_code == 200
@@ -309,7 +299,8 @@ def test_submit_invalid(page, mocker, view, form, viewname):
     response = json.loads(spy.spy_return.content)
     error_message = models.ModelChoiceField.default_error_messages['invalid_choice']
     assert response == {'model_choice': [error_message]}
-    expect(page.locator('[role="group"] ul.dj-errorlist > li.dj-placeholder')).to_have_text(str(error_message))
+    placeholder = page.locator('[role="group"] ul.dj-errorlist > li.dj-placeholder')
+    expect(placeholder).to_have_text(str(error_message))
     initial_opinion.tenant = 1  # reset to initial tenant
     initial_opinion.save(update_fields=['tenant'])
 
@@ -317,23 +308,23 @@ def test_submit_invalid(page, mocker, view, form, viewname):
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['selectize0', 'selectize2', 'selectize4'])
 def test_reset_selectize(page, view, form, viewname):
-    select_element = page.query_selector('django-formset select[is="django-selectize"]')
-    assert select_element is not None
+    select_element = page.locator('django-formset select[is="django-selectize"]')
+    expect(select_element).to_be_visible()
     initial_value = select_element.evaluate('elem => elem.value')
-    input_element = page.query_selector('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
-    assert input_element.is_visible()
+    input_element = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
+    expect(input_element).to_be_visible()
     if form.name in ['selection', 'static_selection']:
         input_element.click()
-        dropdown_element = page.query_selector('django-formset .shadow-wrapper .ts-wrapper .ts-dropdown.single')
-        assert dropdown_element.is_visible()
-        page.wait_for_selector('div[data-selectable]:nth-child(7)').click()
+        dropdown_element = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-dropdown.single')
+        expect(dropdown_element).to_be_visible()
+        page.locator('div[data-selectable]').nth(6).click()
     else:
         input_element.focus()
         page.keyboard.press('Backspace')
     input_element.evaluate('elem => elem.blur()')
     value = select_element.evaluate('elem => elem.value')
     assert value != initial_value
-    page.wait_for_selector('django-formset').evaluate('elem => elem.reset()')
+    page.locator('django-formset').evaluate('elem => elem.reset()')
     value = select_element.evaluate('elem => elem.value')
     assert value == initial_value
 
@@ -341,22 +332,13 @@ def test_reset_selectize(page, view, form, viewname):
 @pytest.mark.urls(__name__)
 @pytest.mark.parametrize('viewname', ['selectize0'])
 def test_touch_selectize(page, form, viewname):
-    field_group = page.query_selector('django-formset [role="group"]')
-    assert 'dj-untouched' in field_group.get_attribute('class')
-    assert 'dj-pristine' in field_group.get_attribute('class')
-    assert 'dj-touched' not in field_group.get_attribute('class')
-    assert 'dj-dirty' not in field_group.get_attribute('class')
-    placeholder = page.query_selector('django-formset ul.dj-errorlist > li.dj-placeholder')
-    assert placeholder.inner_text() == ''
-    input_element = page.query_selector('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
-    assert input_element is not None
+    field_group = page.locator('django-formset [role="group"]')
+    expect(field_group).to_have_class('dj-untouched dj-pristine')
+    placeholder = page.locator('django-formset ul.dj-errorlist > li.dj-placeholder')
+    expect(placeholder).to_have_text('')
+    input_element = page.locator('django-formset .shadow-wrapper .ts-wrapper .ts-control input[type="select-one"]')
+    expect(input_element).to_be_visible()
     input_element.focus()
-    assert 'dj-touched' in field_group.get_attribute('class')
-    assert 'dj-pristine' in field_group.get_attribute('class')
-    assert 'dj-untouched' not in field_group.get_attribute('class')
-    assert 'dj-dirty' not in field_group.get_attribute('class')
-    page.wait_for_selector('django-formset').evaluate('elem => elem.reset()')
-    assert 'dj-untouched' in field_group.get_attribute('class')
-    assert 'dj-pristine' in field_group.get_attribute('class')
-    assert 'dj-touched' not in field_group.get_attribute('class')
-    assert 'dj-dirty' not in field_group.get_attribute('class')
+    expect(field_group).to_have_class('dj-pristine dj-touched')
+    page.locator('django-formset').evaluate('elem => elem.reset()')
+    expect(field_group).to_have_class('dj-pristine dj-untouched')